Create data validation

+++ server/ChelTracker/Controllers/OpponentController.cs
@@ -29,6 +29,11 @@ namespace ChelTracker.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
+
                 var opponents = await _opponentRepository.GetAllUsersOpponentsAsync(userId);
 
                 var opponentsDto = opponents.Select(o => o.ToOpponentDto());
@@ -49,6 +54,11 @@ namespace ChelTracker.Controllers
         [HttpGet("{opponentId}")]
         public async Task<IActionResult> GetById([FromRoute] int opponentId)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var opponent = await _opponentRepository.GetOpponentByIdAsync(opponentId);
 
             if (opponent == null)
@@ -62,16 +72,26 @@ namespace ChelTracker.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateOpponentRequestDto opponentDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var opponentModel = opponentDto.ToOpponentFromCreateDto();
             await _opponentRepository.CreateOpponentAsync(opponentModel);
             return CreatedAtAction(nameof(GetById), new { opponentId = opponentModel.OpponentId }, opponentModel.ToOpponentDto());
         }
 
         [HttpPut]
-        [Route("{opponentId}")]
+        [Route("{opponentId:int}")]
 
         public async Task<IActionResult> Update([FromRoute] int opponentId, [FromBody] UpdateOpponentRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var opponentModel = await _opponentRepository.UpdateOpponentAsync(opponentId, updateDto);
 
             if (opponentModel == null)
@@ -83,9 +103,14 @@ namespace ChelTracker.Controllers
         }
 
         [HttpDelete]
-        [Route("{opponentId}")]
+        [Route("{opponentId:int}")]
         public async Task<IActionResult> Delete([FromRoute] int opponentId)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var opponentModel = await _opponentRepository.DeleteOpponentAsync(opponentId);
 
             if (opponentModel == null)
